+++ database/migrations/2016_01_01_11_create_user_reservation_table.php
@@ -25,14 +25,14 @@ class CreateUserReservationTable extends Migration
             $table->boolean('allow_projector')->nullable();
             $table->integer('request_plug')->unsigned();
             $table->integer('allow_plug')->unsigned();
-            $table->integer('student_id')->unsigned();
+            $table->bigInteger('student_id')->unsigned();
             $table->integer('div_id')->unsigned()->nullable();
             $table->string('other_div')->nullable();
             $table->integer('act_id')->unsigned()->nullable();
             $table->string('other_act')->nullable();
             $table->integer('request_room_id')->unsigned();
             $table->integer('allow_room_id')->unsigned()->nullable();
-            $table->integer('approver_id')->unsigned()->nullable();
+            $table->bigInteger('approver_id')->unsigned()->nullable();
             $table->string('reason_if_not_approve')->nullable();
             $table->timestamp('create_at');
             $table->timestamp('approve_at')->nullable();
